chore: remove old navigator pane
