+++ src/components/organisms/NavigatorPane/NavSectionRenderer/NavSectionItem/styled.tsx
