+++ src/components/organisms/NavigatorPane/NavSectionRenderer/styled.tsx
