+++ src/components/organisms/NavigatorPane/NavigatorPane.styled.tsx
