--- src/components/organisms/NavigatorPane/components/NavigatorContentSubTitle/index.ts
@@ -1,2 +1,0 @@
-export type {NavigatorContentSubTitleProps} from './NavigatorContentSubTitle';
-export {default} from './NavigatorContentSubTitle';
