--- src/components/organisms/NavigatorPane/components/NavigatorContentTitle/index.ts
@@ -1,2 +1,0 @@
-export type {NavigatorContentTitleProps} from './NavigatorContentTitle';
-export {default} from './NavigatorContentTitle';
