--- src/components/organisms/NavigatorPane/components/SectionCol.tsx
@@ -1,10 +1,0 @@
-import {Col} from 'antd';
-import styled from 'styled-components';
-
-const SectionCol = styled(Col)`
-  width: 100%;
-  margin: 0;
-  padding: 0;
-`;
-
-export default SectionCol;
