--- src/components/organisms/NavigatorPane/constants.ts
@@ -1,1 +1,0 @@
-export const ALL_NAMESPACES = '- all -';
