+++ src/components/organisms/index.ts
@@ -3,7 +3,6 @@ export {default as DiffModal} from './DiffModal';
 export {default as StartupModal} from './StartupModal';
 export {default as FileTreePane} from './FileTreePane';
 export {default as MessageBox} from './MessageBox';
-export {default as NavigatorPane} from './NavigatorPane';
 export {default as PageFooter} from './PageFooter';
 export {default as PageHeader} from './PageHeader';
 export {default as SettingsDrawer} from './SettingsDrawer';
@@ -11,3 +10,4 @@ export {default as HotKeysHandler} from './HotKeysHandler';
 export {default as PaneManager} from './PaneManager';
 export {default as NewResourceWizard} from './NewResourceWizard';
 export {default as RenameResourceModal} from './RenameResourceModal';
+export {default as NavigatorPane} from './NavigatorPane';
